♻️  Move '_process_lines_with_discrepancies' to odoo.sync model to use it in globally
