+++ som_sync_openerp/models/payment_order.py
@@ -1,7 +1,6 @@
 #  -*- coding: utf-8 -*-
 from osv import osv
 from service.security import Sudo
-import json
 
 
 class PaymentOrder(osv.osv):
@@ -74,37 +73,6 @@ class PaymentOrder(osv.osv):
         }
         return mapping.get((is_grouped, is_refund))
     # ----------------------------------
-
-    def _get_total_amount_difference(self, inv_read_sync_record):
-        """
-        This method is used to get the total amount difference between Odoo and ERP
-        inv_read_sync_record format expected:
-        {
-            'id': 1,
-            'res_id': 20,
-            'odoo_id': 57,
-            'odoo_last_update_result': '{...}'
-        }
-        """
-        if not inv_read_sync_record or 'odoo_last_update_result' not in inv_read_sync_record:
-            return 0, None
-        odoo_last_update_result = inv_read_sync_record['odoo_last_update_result']
-        if not odoo_last_update_result:
-            return 0, None
-        if not isinstance(odoo_last_update_result, dict):
-            try:
-                odoo_last_update_result = json.loads(odoo_last_update_result)
-            except Exception:
-                return 0
-        if 'data' in odoo_last_update_result and 'metadata' in odoo_last_update_result['data'] \
-                and isinstance(odoo_last_update_result['data']['metadata'], list) \
-                and len(odoo_last_update_result['data']['metadata']) > 0 \
-                and 'pnt_amount_total_erp_difference' in odoo_last_update_result['data']['metadata'][0]:  # noqa: E501
-            discrepancy = (
-                odoo_last_update_result['data']['metadata'][0]['pnt_amount_total_erp_difference'])
-            if discrepancy:
-                return discrepancy, odoo_last_update_result['data']['metadata'][0]['move_type']
-        return 0, None
 
     def _is_order_grouped_invoices(self, cr, uid, payment_order):
         for line in payment_order.line_ids:
@@ -192,7 +160,8 @@ class PaymentOrder(osv.osv):
             cr, uid, inv_sync_with_diff_ids, ['res_id', 'odoo_id', 'odoo_last_update_result'])
         for inv_read_sync_record in inv_read_sync_records:
             # we get the amount difference from the last synchronization
-            amount_difference, move_type = self._get_total_amount_difference(inv_read_sync_record)
+            amount_difference, move_type = sync_obj._get_total_amount_difference(
+                inv_read_sync_record)
             if amount_difference == 0:
                 continue
